+++ python/h3ronpy/pandas/vector.py
@@ -26,6 +26,7 @@ def _geoseries_from_wkb(func, doc: Optional[str] = None, name: Optional[str] = N
 
 
 cells_to_coordinates = _wrap(_av.cells_to_coordinates, ret_type=pd.DataFrame)
+coordinates_to_cells = _wrap(_av.coordinates_to_cells, ret_type=pd.Series)
 cells_bounds = _av.cells_bounds
 cells_bounds_arrays = _wrap(_av.cells_bounds_arrays, ret_type=pd.DataFrame)
 cells_to_wkb_polygons = _wrap(_av.cells_to_wkb_polygons, ret_type=pd.Series)
@@ -113,6 +114,7 @@ def geodataframe_to_cells(
 
 __all__ = [
     cells_to_coordinates.__name__,
+    coordinates_to_cells.__name__,
     cells_bounds.__name__,
     cells_bounds_arrays.__name__,
     cells_to_wkb_polygons.__name__,